Add API to verify jvalue against schema
:Release Notes:
Add C API jvalue_check_schema
Add CPP API JValidator::isValid
Remove a few low-level public C API

:Detailed Notes:
Add APIs to validate jvalue against schema

:Testing Performed:
Internal unit tests were extended to cover new functional
Internal unit tests were successully passed. Tests were run with
control of memory errors

:QA Notes:

:Issues Addressed:
[GF-2836] Validation of json objects against schemas in pbnjson

Open-webOS-DCO-1.0-Signed-off-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>

Reviewed-on: https://g2g.palm.com/2854
Tested-by: Alekseyev Oleksandr <alekseyev.oleksandr@lge.com>

+++ src/test/c_api/TestSchema.cpp
@@ -169,7 +169,7 @@ static QFile* map(QString path, raw_buffer &fData)
 		goto map_failed;
 	}
 
-	if (f->size() > std::numeric_limits<typeof(fData.m_len)>::max()) {
+	if (static_cast<size_t>(f->size()) > std::numeric_limits<typeof(fData.m_len)>::max()) {
 		message = "is too big";
 		goto map_failed;
 	}
@@ -243,7 +243,7 @@ static JSchemaResolutionResult simpleResolver(JSchemaResolverRef resolver, jsche
 	}
 
 	raw_buffer readSchema;
-	if (resourceData.size() > std::numeric_limits<typeof(readSchema.m_len)>::max()) {
+	if (static_cast<size_t>(resourceData.size()) > std::numeric_limits<typeof(readSchema.m_len)>::max()) {
 		qWarning() << "Schema" << lookupPath.absoluteFilePath() << "is too big";
 		return SCHEMA_GENERIC_ERROR;
 	}
