+++ src/test/cxx_api/TestSchema.cpp
@@ -184,7 +184,7 @@ static QFile* map(QString path, raw_buffer &fData)
 		goto map_failed;
 	}
 
-	if (f->size() > std::numeric_limits<typeof(fData.m_len)>::max()) {
+	if (static_cast<size_t>(f->size()) > std::numeric_limits<typeof(fData.m_len)>::max()) {
 		message = "is too big";
 		goto map_failed;
 	}
